requirements.txt
pandas~=2.2.3
torch~=2.5.1+cu124
transformers~=4.45.2
torchvision~=0.20.1+cu124
pillow~=11.0.0